requirements.txt
numpy
opencv-python
pandas
scikit-learn
pillow
seaborn
matplotlib
gradio
tensorflow-cpu==2.12.0
